libs/sdk/python/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "src/buildkite_sdk" ]

[tool.coverage.report]
exclude_lines = ['if TYPE_CHECKING:']
show_missing = true

[tool.pytest.ini_options]
addopts = "--cov --cov-report html:'../../../coverage/libs/sdk/python/html' --cov-report xml:'../../../coverage/libs/sdk/python/coverage.xml' --html='../../../reports/libs/sdk/python/unittests/html/index.html' --junitxml='../../../reports/libs/sdk/python/unittests/junit.xml'"

[project]
name = "buildkite-sdk"
version = "1.0.0"
description = "Automatically generated by Nx."
requires-python = ">=3.9,<4"
readme = 'README.md'
dependencies = [
  "pyyaml"
]

[tool.hatch.build.targets.wheel]
packages = ["src/buildkite_sdk"]

[tool.hatch.metadata]
allow-direct-references = true

[dependency-groups]
dev = [
  "autopep8>=2.3.1",
  "ruff>=0.8.2",
  "pytest>=8.3.4",
  "pytest-sugar>=1.0.0",
  "pytest-cov>=6.0.0",
  "pytest-html>=4.1.1",
  "typing-extensions>=4.12.2"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
exclude = [
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "dist",
]

line-length = 88
indent-width = 4

[tool.ruff.lint]
select = [
  # pycodestyle
  "E",
  # Pyflakes
  "F",
  # pyupgrade
  "UP",
  # flake8-bugbear
  "B",
  # flake8-simplify
  "SIM",
  # isort
  "I",
]
ignore = []

fixable = ["ALL"]
unfixable = []

[tool.uv]
package = true
